refactor: NSImageExtension
